requirements.txt
firebase_functions==0.1.0
functions-framework>=3.0.0
firebase-admin>=6.0.0
python-telegram-bot==13.7.0
requests>=2.20.0